docs: Separate Drawer to new page under Modal
