+++ packages/dnb-design-system-portal/src/docs/uilib/components/modal.md
@@ -1,5 +1,5 @@
 ---
-title: 'Modal / Drawer'
+title: 'Modal'
 description: 'Modal dialogs appear on top of the main content changing the mode of the system into a special mode requiring user interaction'
 order: 11
 showTabs: true
