+++ packages/dnb-design-system-portal/src/docs/uilib/components/modal/drawer.md
@@ -1,0 +1,10 @@
+---
+title: 'Drawer'
+showTabs: true
+---
+
+import DrawerInfo from 'Docs/uilib/components/modal/drawer/info'
+import DrawerDemos from 'Docs/uilib/components/modal/drawer/demos'
+
+<DrawerInfo />
+<DrawerDemos />
