+++ packages/dnb-design-system-portal/src/docs/uilib/components/modal/drawer/demos.md
@@ -1,0 +1,28 @@
+---
+showTabs: true
+---
+
+import {
+ModalExampleDrawerHeader,
+ModalExampleDrawerBasic,
+ModalExampleCloseByCallback,
+DrawerRealLifeExample,
+} from 'Docs/uilib/components/modal/drawer/Examples'
+
+## Demos
+
+### Basic drawer
+
+<ModalExampleDrawerBasic />
+
+### Drawer with custom content
+
+<DrawerRealLifeExample />
+
+### Drawer with navigation
+
+<ModalExampleDrawerHeader />
+
+### Close Drawer by callback method
+
+<ModalExampleCloseByCallback />
