+++ packages/dnb-design-system-portal/src/docs/uilib/components/modal/drawer/events.md
@@ -1,0 +1,7 @@
+---
+showTabs: true
+---
+
+## Events
+
+Find the properties in [Modal Events](/uilib/components/modal/events).
