+++ packages/dnb-design-system-portal/src/docs/uilib/components/modal/drawer/properties.md
@@ -1,0 +1,7 @@
+---
+showTabs: true
+---
+
+## Drawer Properties
+
+Find the properties in [Modal Properties](/uilib/components/modal/properties).
